% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_alignR.R
\name{align_rasters}
\alias{align_rasters}
\title{Align a List of SpatRaster Objects to a Template Raster}
\usage{
align_rasters(raster_list, template, resample_method = "bilinear")
}
\arguments{
\item{raster_list}{A list of \code{SpatRaster} objects to be checked and aligned.}

\item{template}{A \code{SpatRaster} object used as the reference for alignment.}

\item{resample_method}{Character. The resampling method to use if alignment is needed.
Options are \code{"bilinear"} (default), \code{"near"} (nearest neighbor), \code{"cubic"}, \code{"cubicspline"}, etc.}
}
\value{
A list of \code{SpatRaster} objects, where each raster is either unchanged (if already aligned)
or resampled to match the template.
}
\description{
This function iterates over a list of \code{SpatRaster} objects, checks if each raster
is aligned with a given template raster (in terms of extent, resolution, projection),
and if not, resamples it to match the template.
}
\details{
The function checks alignment using \code{terra::compareGeom()}, which verifies if the
extent, resolution, and coordinate reference system (CRS) match between the raster
and the template. If differences are detected, \code{terra::resample()} is used to align the raster.
}
\examples{
library(terra)
r1 <- rast(ncol=10, nrow=10)
r2 <- rast(ncol=12, nrow=12)  # Different resolution
template <- rast(ncol=10, nrow=10)

rasters <- list(r1, r2)
aligned_rasters <- align_rasters(rasters, template, resample_method = "bilinear")

}
