% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookiecutteR.R
\name{cookie_cutter}
\alias{cookie_cutter}
\title{Parallel crop, mask, and export/return for raster x polygon combos}
\usage{
cookie_cutter(
  rasters,
  polygons,
  raster_names = NULL,
  outpath = NULL,
  mode = "return",
  overwriteRaster = TRUE
)
}
\arguments{
\item{rasters}{A list of SpatRaster objects.}

\item{polygons}{A list of polygon objects (either sf or SpatVector).}

\item{raster_names}{Optional vector of names for rasters.}

\item{outpath}{Optional string directory path (used if mode = "export" or "both").}

\item{mode}{One of "return", "export", or "both".}

\item{overwriteRaster}{Logical. Whether to overwrite existing raster files on disk. Default is TRUE.}
}
\value{
Named list of processed rasters (if mode is "return" or "both")
}
\description{
Parallel crop, mask, and export/return for raster x polygon combos
}
