% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nochh_mskR.R
\name{no_chmsk}
\alias{no_chmsk}
\title{Create a Mask for No-Change or Change Across Multiple spatRaster Layers or Objects}
\usage{
no_chmsk(..., type = "nochange")
}
\arguments{
\item{...}{Multiple spatRaster objects or a single multiband spatRaster.}

\item{type}{A character string, either "nochange" to create a no-change mask
or "change" to create a change mask. Default is "nochange".}
}
\value{
A spatRaster object with 1 indicating no-change or change pixels
and NA for the rest.
}
\description{
This function takes a multiband spatRaster or multiple spatRaster objects
and produces a mask indicating pixels that remained the same or changed
across all layers/objects.
}
\examples{
# Given r1, r2, and r3 are spatRaster objects
# Create a no-change mask:
mask <- no_chmsk(r1, r2, r3, type = "nochange")
# Create a change mask:
mask <- no_chmsk(r1, r2, r3, type = "change")

# Given multi_band_raster is a multiband spatRaster
# Create a no-change mask:
mask <- no_chmsk(multi_band_raster, type = "nochange")
}
