% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizR.R
\name{normalize_raster}
\alias{normalize_raster}
\title{Normalize a Raster}
\usage{
normalize_raster(raster)
}
\arguments{
\item{raster}{A \code{SpatRaster} object to be normalized.}
}
\value{
A normalized \code{SpatRaster} object.
}
\description{
This function normalizes a raster by rescaling its values to the range \link{0, 1}.
}
\examples{
r <- terra::rast(matrix(1:100, ncol = 10))
normalize_raster(r)
}
