% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KS_helpers.R
\name{plot_ks_bars}
\alias{plot_ks_bars}
\title{Top-K services by KS D for a chosen variable}
\usage{
plot_ks_bars(ks_aug, var_id, top_k = 10, palette = "C")
}
\description{
Top-K services by KS D for a chosen variable
}
