% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KS_helpers.R
\name{plot_ks_mountain}
\alias{plot_ks_mountain}
\title{Mountain plot (CDF difference) for one service × variable}
\usage{
plot_ks_mountain(
  hotspots_df,
  inverse_df,
  service_id,
  var_id,
  transform = c("identity", "log1p")[2],
  n_grid = 512
)
}
\description{
Mountain plot (CDF difference) for one service × variable
}
