% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pix_area.R
\name{sample_pixels}
\alias{sample_pixels}
\title{Sample Pixels from a Raster with Constraints}
\usage{
sample_pixels(raster, target_area_ha, mask = NULL, min_distance = NULL)
}
\arguments{
\item{raster}{A \code{SpatRaster} object. The raster from which pixels will be sampled.}

\item{target_area_ha}{Numeric. The target area in hectares to be sampled, converted to pixel count.}

\item{mask}{A \code{SpatRaster} object (optional). A mask defining valid sampling regions (1 for valid, NA for invalid).}

\item{min_distance}{Numeric (optional). Minimum distance (in pixels) from raster edges or NA areas.}
}
\value{
A \code{SpatRaster} object with the sampled pixels.
}
\description{
This function samples a specified number of pixels from a raster while ensuring
constraints like a minimum distance from raster edges or masked areas.
}
\examples{
# Example usage:
library(terra)
r <- rast(nrows=100, ncols=100, res=30, vals=runif(10000, 0, 1))
mask <- r
values(mask) <- ifelse(runif(ncell(mask)) > 0.5, 1, NA)
sample <- sample_pixels(raster = r, target_area_ha = 50, mask = mask, min_distance = 5)
plot(sample)
}
